Test for desktop support

+++ install.sh
@@ -9,68 +9,109 @@
 
 TargetDir="${HOME}"
 RunDir="$(dirname $0)"
-Confile="$HOME/.config/KaOS-restore.log"
+yad=$(which yad)
+DesktopList="Plasma Enlightenment Gnome Mate Xfce Lxde Lxqt Openbox Kde4"
 
-echo "">>${Confile} || exit 1
+# First verify we can dialog thru X
+if [ -z "$(which yad)" ] && [ -z $(which kdialog) ] && [ -z $(which zenity) ];then
+        echo "Sorry, this script need kdialog or yad or zenity to run."
+        exit 1
+fi
 
 # For now this just verify we are on a KaOS system but it shoud next identify which distrib it is
 if [ -f /etc/lsb-release ]
 then
         Distrib=$(grep "DISTRIB_ID=" /etc/lsb-release | cut -d '=' -f 2)
-        [ "${Distrib}" = "KaOS" ] && [  -d ${RunDir}/Distrib/KaOS ] && [ -d ${RunDir}/Desktops/Plasma ] && \
-        _echo='kdialog --title "KaOS Dotfiles Restore" --msgbox' || exit 1
+        if [ "${Distrib}" = "KaOS" ] && [  -d ${RunDir}/Distrib/KaOS ] && [ -d ${RunDir}/Desktops/Plasma ] ;then
+                _echo='kdialog --title "KaOS Dotfiles Restore" --msgbox'
+                _Desktop=Plasma
+        elif [ "${Distrib}" = "Manjaro" ] && [  -d ${RunDir}/Distrib/Manjaro ];then
+                _echo="${yad} --title='${Distrib} Dotfiles Restore' --text"
+                _Desktop=$(AskDesktop)
+        fi
 else 
-        exit 1  
+        echo "/etc/lsb-release file not found... Exiting..."
+        exit 1
 fi
 
+# Create some memory file
+Confile="$HOME/.config/${Distrib}-restore.log"
+if [ ! -f "${Confile}" ];then
+        echo "">>${Confile}
+fi
+
+# Error message
 _error() {
         kdialog --error "$1"
 }
 
+# Error message and exit
 _errorexit() {
         kdialog --error "$1" && exit 0;
 }
 
-_postinstall() {
+# Make copy with .ori backup
+cpb() { cp $@{,.ori} ;}
+
 # Once identified it should run the appropriate distrib specific install-script
-case $1 in 
-        KaOS)DistribDir="${RunDir}/Distrib/$1"
+_postinstall() {
+case "$1" in 
+        "KaOS")DistribDir="${RunDir}/Distrib/$1"
                 [ -f ${DistribDir}/$1-post-install.sh ] && chmod +x ${DistribDir}/$1-post-install.sh && $1_postinstall || exit 1;;
-        Manjaro) echo "Doing Manja stuff";;
+        "Manjaro") echo "Doing Manja stuff... Soon...";;
+        "Debian") echo "First install pacapt... Soon...";;
+        "Ubuntu") echo "First install pacapt... Soon...";;
         *) echo "Distribution not (yet?) supported" && return 1;;
 esac
 }
-export _postinstall
 
 KaOS_postinstall() {
 # Not sure how it should be launch...
 [ "$?" = "0" ] && bash -x ${DistribDir}/KaOS-post-install.sh || return 1
 }
 
+#Ask for used desktop
+AskDesktop() {
+        for d in ${DesktopList};do echo FALSE ${d};done | ${yad} --text="Which Desktop do you use?\n" \
+        --list --radiolist --column="Installed" --column="Desktop" \
+        --listen --print-column=2
+}
+
 ## Bash files
 _bashinstall() {
 # For now only restore few dotfiles but then should ask for lib to use etc...
-        read
         # Save existing files
         if [ -f $HOME/.bashrc ];then cpb $HOME/.bashrc;fi
         if [ -f $HOME/.bash_aliases ];then cpb $HOME/.bash_aliases;fi
         if [ -f $HOME/.bash_functions ];then cpb $HOME/.bash_functions;fi
         
-        cat bashrc ${1}/${1}.bashrc > ${TargetDir}/.bashrc || return 1	# Copy bashrc
-        cat bash_aliases ${1}/${1}_aliases perso/perso_aliases > ${TargetDir}/.bash_aliases || return 1    # Create .bash_aliases
-        cat bash_functions ${1}/${1}_functions perso/perso_functions > ${TargetDir}/.bash_functions || return 1    # Create .bash_functions
+        cat bashrc Distrib/${1}/${1}.bashrc > ${TargetDir}/.bashrc || return 1	# Copy bashrc
+        if [ -f ${RunDir}/Desktops/${_Desktop}/bash_aliases ];then
+                cat bash_aliases ${1}/${1}_aliases perso/perso_aliases Desktops/${_Desktop}/bash_aliases > ${TargetDir}/.bash_aliases
+        else
+                cat bash_aliases ${1}/${1}_aliases perso/perso_aliases > ${TargetDir}/.bash_aliases
+        fi
+        if [ -f ${RunDir}/Desktops/${_Desktop}/bash_functions ];then
+                cat bash_functions ${1}/${1}_functions perso/perso_functions Desktops/${_Desktop}/bash_functions > ${TargetDir}/.bash_functions
+        else     
+                cat bash_functions ${1}/${1}_functions perso/perso_functions > ${TargetDir}/.bash_functions || return 1    # Create .bash_functions
+        fi
         
         # Want to boost bash?
         # Awesome fonts?
-        if [ -d fonts ];then ${yad} --text='Do you want to install extras and awesome fonts?'
-        if  [ "$?" = "0" ];then ${_terminal} -e "./install.sh";fi
+        if [ -d "${RunDir}/fonts" ];then 
+                ${yad} --text='Do you want to install extras and awesome fonts?'
+        fi
+        if  [ "$?" = "0" ];then 
+                ${_terminal} -e "./install.sh"
+        fi
         
         # TODO: better getopt?
         # Liquidpromt?
         # powerline?
         
         if [ -f "${1}/post-${1}-install.sh" ]   # If exists, run post-install script
-        then ./${1}/post-${1}-install.sh || return 1
+        then echo -e "Running post install script now..." && ./${1}/post-${1}-install.sh || return 1
     fi
     echo -e 'Dot files installed!\nBye!'
     return 0
